[pyyamproto]: Makefiles for build protobufs into an egg
This mainly consists of a makefile to turn the proto files into generated
python and create an egg from them. A current version of the python files are
checked in to make it so protoc isn't needed on the machine to use the
projects unless you modify the protocol definitions.