[PYYAM]: Add ZMQ Process encapsulation and message handler
To encapsulate all the ZMQ logic into a separate process that can run
independently of the server we use ZMQProcess. The ZMQ stream requires a
message handler. We a have a message handler base class which can be used to
impelment message handling in the server and client without doing anything too
complex. The message handlers know how to reassembyle the ZMQ message to look
at the type and pass it along to the right handler.